+++ app/home/forecast/page.tsx
@@ -1,6 +1,7 @@
 import { fetchLighthouses, getWeather } from "@/app/lib/data";
 import Header from "@/app/ui/forecast/dailyTable";
 import DayCard from "@/app/ui/forecast/dayCard";
+import TableRow from "@/app/ui/forecast/hourlyTable";
 import Map from '@/app/ui/home/map';
 
 export default async function Page() {
@@ -10,12 +11,21 @@ export default async function Page() {
       <h1>Current Weather</h1>
       {lighthouses.map(async(lighthouse) => {
         const { current, daily, hourly } = await getWeather(lighthouse.latitude, lighthouse.longitude, 'Europe/Dublin' );
-        console.log("Current....",current);
-        console.log(lighthouse.name);
+        // console.log("Current....",current);
+        // console.log(lighthouse.name);
 
         return (
           <>
             <h1>{ lighthouse.name }</h1>
+            {/* <div className="bg-white-700 mx-auto my-5 w-[40%] h-[40%]">
+                    <Map 
+                    id={lighthouse.id} 
+                    name={lighthouse.name} 
+                    latitude={lighthouse.latitude} 
+                    longitude={lighthouse.longitude} 
+                    />
+            </div> */}
+            
             <Header
               currentTemp={current?.currentTemp}
               highTemp={current?.highTemp}