+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Irish Beacon Watch Changelog
 
+## verson 0.2.0
+Added linegraph to Historical page to show wind metrics
+Added linegraph to Marine page to show waveheight metrics
+
 ## verson 0.1.1
 Added map to home page
 Added weather api data forecast page (graphs and tables are still to be decided on)
 Rendered weather data on forecast page
 Implemented function to retrieve lighthouse data from database onto map page
+Added Historical API call to data file
+Added test historical weather table to Historical page
 
 ## verson 0.1.0
 Clone example boiler plate from nextjs official web page https://nextjs.org/learn/dashboard-app/getting-started